a64.sh:  line 42: tar -xvf: command not found [FIXED]
a64.sh script failed to extract the downloaded JDK. I am not familiar with shell scripting. But, I have a lot of programming experience. So, from my experience I think that when one command assigned to the variable as a value, the shell assumes that command as a normal data(including spaces between "tar" and "-xvf") and not as a command. So, you should directly write the command instead of assigning in one variable and then putting that variable or there may be something other ways available(I don't know).
Though there is nothing extraction happened, a64.sh script skipped that and went to the next operation i.e. Setting-up environment variable %JAVA_HOME%... and Cleaning up temporary files.